Rewrote /analytics/municipalities with saving stats as json file
